+++ CHANGELOG.md
@@ -2,10 +2,30 @@
 
 All notable changes to this project will be documented in this file.
 
+## 0.3.4 (2023-08-02)
+
+### Added
+
+- added start() function
+- added stop() function
+- added frequency() function
+- added precision() function
+- added x86_64_processor_id() function
+- added x86_64_measure_frequency() function
+
+### Deprecated
+
+- deprecated tick_counter_start() function
+- deprecated tick_counter_stop() function
+- deprecated tick_counter_frequency() function
+- deprecated tick_counter_precision_nanoseconds() function
+- deprecated x86_64_tick_counter_processor_id() function
+- deprecated x86_64_measure_tick_counter_frequency() function
+
 ## 0.3.3 (2023-07-28)
 
-* added categories to metadata
+- added categories to metadata
 
 ## 0.3.2 (2023-07-12)
 
-* updated documentation
+- updated documentation